Predict virtual hands
